agents modeling/Agents/Forms Agent/requirements.txt
# Environment management
python-dotenv

# Templating
jinja2

# Pinecone
pinecone-client
langchain-pinecone

# LangChain document loading (community-maintained)
langchain-community

# Google Gemini Embeddings
google-generativeai

# Web server
flask

# LangGraph (for agent orchestration)
langgraph

# Type support (for static type checking, optional but helps with Annotated, TypedDict, etc.)
typing-extensions

datetime
